+++ next.config.ts
@@ -1,7 +1,6 @@
 const path = require("path");
 const withPWA = require("next-pwa")({
   dest: "public",
-  disable: process.env.NODE_ENV === "development",
 });
 
 const withSvgr = require("next-svgr");
